modules/report_analyzer/requirements.txt
Flask>=2.0.0
groq>=0.18.0
pdfplumber>=0.11.5
python-dotenv